pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["benben*"]

[project]
name = "benben"
version = "0.1.0"
description = "Ben的本本"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "flask>=3.0,<4",
    "openai==1.35.0",
    "python-dotenv==1.0.1",
    "gunicorn==21.2.0",
    "requests==2.31.0",
    "numpy>=1.26.4,<2.0.0",
    "faiss-cpu>=1.10,<1.14",
    "pyyaml==6.0.1",
    "ruamel.yaml==0.18.15",
    "oss2>=2.18,<3",
    "markdown-it-py>=3,<4",
    "mdit-py-plugins>=0.4,<0.5",
    "beautifulsoup4==4.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4,<9",
]
